+++ app/src/main/java/com/github/bumblebee202111/doubean/ui/component/FullScreenCenteredContent.kt
@@ -1,0 +1,24 @@
+package com.github.bumblebee202111.doubean.ui.component
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxScope
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun FullScreenCenteredContent(
+    paddingValues: PaddingValues,
+    content: @Composable BoxScope.() -> Unit,
+) {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(paddingValues),
+        contentAlignment = Alignment.Center,
+        content = content
+    )
+}